chore: Add submodule for glm library
This commit adds the `glm` library as a submodule. The submodule is located in the `vendors/glm` directory and is sourced from the `https://github.com/g-truc/glm.git` repository. This addition is necessary to incorporate the `glm` library into the project and enable its usage for OpenGL rendering operations.